python_backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
flask-cors==4.0.0
google-generativeai==0.8.3
python-docx==1.1.0
pdfminer.six==20231228
PyMuPDF==1.23.26
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
nltk==3.8.1
textstat==0.7.3
python-dotenv==1.0.0
bcrypt==4.1.2
pymongo==4.6.1
Pillow==10.1.0
requests==2.31.0
wordcloud==1.9.2
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
regex==2023.12.25
